requirements.txt
# Webhook server and Cast dependencies
flask==2.3.3
requests==2.31.0
pychromecast==13.0.4
zeroconf==0.112.0
python-dotenv==1.0.0
gunicorn==21.2.0

# RTSP and stream processing
ffmpeg-python==0.2.0